Update OOAD LAB2 - Duck Hunt.

+++ Course/OOAD/LAB2 - DuckHunt/src/main/java/io/billc/duckhunt/Utils/Config.java
@@ -14,7 +14,7 @@ public class Config {
 	public static final Dimension FRAME_SIZE = new Dimension(800, 500);
 	public static final int GAME_PLAY_DURATION = 50;
 	public static final int DUCK_SIZE = 45;
-	public static final float DUCK_FLY_SPEED_FACTOR = 1;
+	public static final float DUCK_FLY_SPEED_FACTOR = (float) 1;
 	public static final int DUCK_LIFESPAN_MILLS = 18000;
 	public static final int DUCK_PER_WAVE = 10;
 	public static final int WAVE_INTERVAL_SECS = 10;
